test(HealthNexus/KeyManager):kc protractor edits
