+++ docs/index.html
@@ -1,1 +1,1 @@
-<!doctype html><html lang="en"><head><meta charset="utf-8"><title>HealthNexusWallet</title><base href="KeyManager"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" type="image/x-icon" href="favicon.ico"><link href="styles.e2e7e26a36739edf69d7.bundle.css" rel="stylesheet"/></head><body><app-root></app-root><script type="text/javascript" src="inline.318b50c57b4eba3d437b.bundle.js"></script><script type="text/javascript" src="polyfills.f8e21e048b5014eb40c7.bundle.js"></script><script type="text/javascript" src="main.2a85b3d524c8630b0f9f.bundle.js"></script></body></html>
+<!doctype html><html lang="en"><head><meta charset="utf-8"><title>HealthNexusWallet</title><base href="KeyManager"><meta name="viewport" content="width=device-width,initial-scale=1"><link rel="icon" type="image/x-icon" href="favicon.ico"><link href="styles.e2e7e26a36739edf69d7.bundle.css" rel="stylesheet"/></head><body><app-root></app-root><script type="text/javascript" src="inline.318b50c57b4eba3d437b.bundle.js"></script><script type="text/javascript" src="polyfills.f8e21e048b5014eb40c7.bundle.js"></script><script type="text/javascript" src="main.f1f9749c04f9153d7131.bundle.js"></script></body></html>
