--- e2e/environment/app.accounts-info.ts
@@ -1,12 +1,0 @@
-// test account information:
-// account #1
-export let account1public = '';
-export let account1private = '';
-export let account1mnemonic = '';
-export let account1password = '';
-
-// account #2
-export let account2public = '';
-export let account2private = '';
-export let accout2mnemonic = '';
-export let account2password = '';
