+++ e2e/environment/app.accountsInfo.ts
@@ -1,0 +1,25 @@
+// // test account information:
+// // account #1
+// export let account1public = '';
+// export let account1private = '';
+// export let account1mnemonic = '';
+// export let account1password = '';
+
+// // account #2
+// export let account2public = '';
+// export let account2private = '';
+// export let accout2mnemonic = '';
+// export let account2password = '';
+
+// test account information:
+// account #0
+export let account1public = '0xA0031fdBD932CBa7B19426FfD8317F4390831985';
+export let account1private = '9b081cac311a8ad6df622f388e5e334790b33f4d862e3210a4d3121caf95fa16';
+export let account1mnemonic = 'lucky tide ritual belt fruit trumpet hurdle elevator jar else lunch morning';
+export let account1password = 'ocVf45ty!!!!-';
+
+// account #2
+export let account2public = '0x3e460451a58D87c065fF1dA8cb59c26502Db37A8';
+export let account2private = 'ab5e67e54aa59a1a9eefe4c2b2a517e6a52c3dc00c5ff43b947cb5cd797213ee';
+export let accout2mnemonic = 'lucky tide ritual belt fruit trumpet hurdle elevator jar else lunch morning';
+export let account2password = 'ocVf45ty!!!!-';
